Refactor communicate prompt to be less strict.
Checking if a script is possible before trying to write it is filtering
out too many possible scripts. Often the AI will say it's not possible,
and then write the script in it's comment anyways, which is baffling
when it happens. It may be worth revisiting this later by adding a
"force" button to the UI that just skips this check, but for now the
goofy scripts this filters out are not really a problem.