Remove unused "investigate" function
Originally it was planned to have Aerome query OpenAI for actions that
could be taken on each folder, for instance, if you're in a folder for a
Rust project then the "build" or "run" action could pop up automatically
as a suggesstion. This is a good idea, but it's being cut from the first
release. Sending every file that's being navigated through to OpenAI
kinda creeps me out, and I'm excited about this project! Finding a way
to do this in a way that respects the users privacy is on the TODO.