Fix error when copying symlinks.
The fs_extra libraries default behaviour when copying symlinks was to
read the file the symlink pointed to and copy it into a new file, with
no way to configure it. Worse still, if it came across a broken symlink
it stopped the file transfer process altogether with an error. A fork of
the repository with a fix to cut/copy just the symlink files directly
is already done but hasn't been pulled into the main repo yet. So we
forked the fork with the fix and we're now relying on that.